+++ src/rbx/mod.rs
@@ -1,4 +1,6 @@
 mod message;
+mod remodel;
 mod universe;
 pub use message::*;
+pub use remodel::*;
 pub use universe::*;
